Create SubscriptionPlans component
